making changes to db layer of courses to include user_id

+++ src/server/db/coursesDB.js
@@ -1,27 +1,28 @@
 const pool = require("./pool.js");
 
-async function findAll() {
-    const {rows} = await pool.query('SELECT * FROM courses');
+async function findAll(user_id) {
+    
+    const {rows} = await pool.query('SELECT * FROM courses WHERE user_id = $1',[user_id]);
     return rows;
 }
 
-async function findById(id) {
-    const {rows} = await pool.query('SELECT * FROM courses WHERE id = $1',[id]);
+async function findById(id,user_id) {
+    const {rows} = await pool.query('SELECT * FROM courses WHERE id = $1 AND user_id=$2',[id,user_id]);
     return rows[0];
 }
 
-async function create(name){
-    const {rows} = await pool.query('INSERT INTO courses(name) VALUES($1) RETURNING *',[name]);
+async function create(name,user_id){
+    const {rows} = await pool.query('INSERT INTO courses(name,user_id) VALUES($1,$2) RETURNING *',[name,user_id]);
     return rows[0];
 }
 
-async function update(name,id){
-    const {rows} = await pool.query('UPDATE courses SET name = $1 WHERE id = $2 RETURNING *',[name,id]);
+async function update(name,id,user_id){
+    const {rows} = await pool.query('UPDATE courses SET name = $1 WHERE id = $2 AND user_id = $3 RETURNING *',[name,id,user_id]);
     return rows[0];
 }
 
-async function remove(id){
-    const {rows} = await pool.query('DELETE FROM courses WHERE id = $1 RETURNING *',[id]);
+async function remove(id,user_id){
+    const {rows} = await pool.query('DELETE FROM courses WHERE id = $1 AND user_id = $2 RETURNING *',[id,user_id]);
     return rows[0];
 }
